Rename http_client to node_client
