+++ chain-signatures/node/src/lib.rs
@@ -2,7 +2,7 @@ pub mod cli;
 pub mod config;
 pub mod contract_updater;
 pub mod gcp;
-pub mod http_client;
+pub mod node_client;
 pub mod indexer;
 pub mod kdf;
 pub mod mesh;
