+++ chain-signatures/node/src/mesh/mod.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use crate::http_client::NodeClient;
+use crate::node_client::NodeClient;
 use crate::protocol::contract::primitives::Participants;
 use crate::protocol::ProtocolState;
 use std::sync::Arc;
